Parse JSON and create questions/options
