+++ config/routes.rb
@@ -1,14 +1,14 @@
 Rails.application.routes.draw do
   root to: 'pages#home'
 
-  resources :forms, only: [:new, :create] do
+  resources :forms, only: [:new, :create, :destroy] do
     get 'twilio', to: 'forms#twilio'
+    post 'recieve', to: 'forms#recieve'
+    post 'start', to: 'forms#start'
+    post 'update', to: 'forms#update'
   end
 
-  get 'forms/create_dev', to: 'forms#create_dev'
-
-  post '/twilio/recieve' => 'twilio#recieve'
-  post '/twilio/start' => 'twilio#start'
+  #get 'forms/create_dev', to: 'forms#create_dev'
 
   #resources :twilio do
     #post 'recieve', to: 'twilio#recieve'
